build: add MSAL single file publish workaround
Add a workaround for a bug in MSAL's loading of native dependencies when
published as a single file bundle.

Using the IncludeAddContentForSelfExtract option will cause all managed
and native dependencies to be extracted on first launch to a temporary
directory, and executed from there.

This only affects Windows as only Windows has support for the MSAL
runtime that requires the native dependencies.